+++ gtd_fact_tree/src/fact.c
@@ -34,62 +34,7 @@ void delete_fact(Fact *fact)
     return;
 }
 
-/**
- * \brief function to check how many parameters are required for fact type
- * \param type fact type
- * \return number of parameters needed
-*/
-uint8_t get_param_count(FactType type)
-{
-    switch (type)
-    {
-    case IstnaryTreeFact:
-    case IsPartiteFact:
-    case VertexCountFact:
-    case MinVertexCountFact:
-    case MaxVertexCountFact:
-    case EdgeCountFact:
-    case MinEdgeCountFact:
-    case MaxEdgeCountFact:
-    case TreeHeightFact:
-    case MinTreeHeightFact:
-    case MaxTreeHeightFact:
-    case HasCycleFact:
-    case HasNoCycleFact:
-    case HasInducedCycleFact:
-    case HasNoInducedCycleFact:
-    case HasMinorCycleFact:
-    case HasNoMinorCycleFact:
-    case HasPathFact:
-    case HasNoPathFact:
-    case HasInducedPathFact:
-    case HasNoInducedPathFact:
-    case HasMinorPathFact:
-    case HasNoMinorPathFact:
-    case HasCliqueFact:
-    case HasNoCliqueFact:
-    case HasMinorCliqueFact:
-    case HasNoMinorCliqueFact:
-        return 1;
-    case IsConnectedFact:
-    case IsTreeFact:
-    case IsPlanarFact:
-    case IsCycleFact:
-    case IsCycleComplementFact:
-    case HasNoCyclesFact:
-        return 0;
-    case HasCompletePartiteFact:
-    case HasNoCompletePartiteFact:
-    case HasInducedCompletePartiteFact:
-    case HasNoInducedCompletePartiteFact:
-    case HasMinorCompletePartiteFact:
-    case HasNoMinorCompletePartiteFact:
-        // handle only bipartite for now
-        return 3;
-    default:
-        return 0;
-    }
-}
+
 
 /**
  * \brief function to check if 2 facts are equal
